Refact: refact 모달버튼 컴포넌트 생성
